+++ platform/ui/src/components/userPreferencesForm/index.js
@@ -1,0 +1,4 @@
+export { UserPreferences } from './UserPreferences.js';
+export { AboutContent } from '../content/aboutContent/AboutContent.js';
+export { UserPreferencesForm } from './UserPreferencesForm.js';
+export { GeneralPreferences } from './GeneralPreferences.js';