+++ platform/viewer/cypress/integration/common/OHIFCornerstoneHotkeys.spec.js
@@ -46,8 +46,8 @@ describe('OHIF Cornerstone Hotkeys', () => {
     // Hotkey I
     cy.get('body').type('I');
     // Visual comparison
-    cy.screenshot();
-    cy.percyCanvasSnapshot('Hotkey I - invert image');
+    cy.screenshot('Hotkey I - Should Invert Image');
+    cy.percyCanvasSnapshot('Hotkey I - Should Invert Image');
   });
 
   it('checks if hotkeys "+", "-" and "=" can zoom in, out and fit to viewport', () => {
@@ -76,8 +76,8 @@ describe('OHIF Cornerstone Hotkeys', () => {
     cy.get('@viewportInfoBottomRight').should('contains.text', 'Zoom: 211%');
 
     // Visual comparison to make sure the 'inverted' image was reset
-    cy.screenshot();
-    cy.percyCanvasSnapshot('Hotkey SPACEBAR - Reset Image');
+    cy.screenshot('Hotkey SPACEBAR - Should Reset Image');
+    cy.percyCanvasSnapshot('Hotkey SPACEBAR - Should Reset Image');
   });
 
   it('uses hotkeys "RightArrow" and "LeftArrow" to navigate between multiple viewports', () => {